requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
openai>=1.45.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
numpy>=1.26.0
pymupdf>=1.24.0
python-docx>=1.1.0
chardet>=5.2.0
python-pptx>=0.6.21
pandas>=2.2.0
